Integrate redux into sign in flow
Added mapStateToProps function to app.js to give it access to state as
well as the current user authentication object. Redirects based on user
signed in/out status. Added user.types file to redux user slice so as
to ensure constancy with variables and naming between files. Updated
reducer and action files with new user types.