Add cart-dropdown component
Added a new component called cart-dropdown, which will be conditionally
rendered in the header component that it was imported into. When the C
artIcon component is clicked by a user, the CartDropdown will appear ov
er the app content, and will display any selected items. Currently is
'hard coded' but will integrate functionality of show/hidden in redux.