Integrate Redux for multiple selected items
Created new cart.utils file which uses a combination of find and map me
thods to determine if an item has already been added to the array of it
ems in the cart. If an item has been added, the quantity in the store i
s increased by 1. If it is not, a new item is added, and the quantity i
s set to 1. Updates types, action, and reducer to allow cartItem state
array to update values properly. Integrated action into collection comp
onents.