Add reselect package, selectors to cart
Added reselect npm package and cart.selectors.js to improve performance
and fully leverage redux functionality via selected slices of state, as
opposed to sending the entire state object and destructuring off of it.
Integrated count selector for cartItems, and updated the cartIcon compo
nent as well as the cartDropdown component. cartItem count appears to b
e working as expected, and will not rerender unless the selected slice
has changed.