Integrate remove item functionality
Added new type, and action in the cart Redux section, to allow the chec
kout page to have the ability to remove items which have been added.
When the wingding X is clicked, the newly created clearItemFromCart fx
is triggered, which filters the cartItems on the state with the same ID
as those in the section which got told to remove. Will return an array
of only those objects which do not match. Function appears to work as
expected.