Add removeItem, addItem functionality
Added left and right arrows to checkout item component. Created new act
ion removeItem in redux, and integrated this, addItem, and removeItemFr
omCart into checkout item, and bound them to onClick functions for the
respective elements. If the X is clicked, the rwo is removed. If the ri
ght facing arrow is clicked, the item quantity increases by one. If the
left facing arrow is clicked, the quantity decreases by 1.