Integrate session persistence
Added redux-persistence package to maintain a user's session and data,
given that we want our users to keep their carts as is in the event the
y refresh the page. Integrated redux-persistence config into store as w
ell as root reducer. Added some items to cart and refreshed and persist
ence seems to be persisten'