Integrate directory state into redux
Currently the directory component is a class component, holding state b
ut for only one property called sections, which is an array of the shop
sections and the associated content for them. Although this should not
be updated by user action at this time, it should make testing of the a
pp easier. Additionally, I was able to write a selector for those secti
ons, which should improve performance should that data increase exponen
tially with more content.